requirements.txt
pandas>=2.0.0
numpy>=1.24.0
jupyter>=1.0.0
ipywidgets>=8.0.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0